% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_matrix.R
\name{create_summary_matrix}
\alias{create_summary_matrix}
\title{Create summary matrix of NPI efficiency}
\usage{
create_summary_matrix(folder, date = Sys.Date())
}
\arguments{
\item{folder}{The folder containing the result of the MCMC run}
}
\description{
Create summary matrix of NPI efficiency
}
