% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_contact_data.R
\name{load_contact_data}
\alias{load_contact_data}
\title{Load contact dataset from specific country}
\usage{
load_contact_data(country)
}
\arguments{
\item{country}{Character string. Name of the country for which you want to
run this analysis}
}
\description{
Load contact dataset from specific country
}
