% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_country_data.R
\name{load_country_data}
\alias{load_country_data}
\title{Load and clean up dataset from specific country}
\usage{
load_country_data(country, dataset_path = NULL)
}
\arguments{
\item{country}{Character string. Name of the country for which you want to
run this analysis}

\item{dataset_path}{Path of the dataset with the epidemiological data for
all countries}
}
\description{
Load and clean up dataset from specific country
}
