% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_epi_data.R
\name{load_epi_data}
\alias{load_epi_data}
\title{Get epidemiological data for 31 European countries}
\usage{
load_epi_data()
}
\description{
Get epidemiological data for 31 European countries
}
\references{
https://www.ecdc.europa.eu/en/publications-data/download-todays-data-geographic-distribution-covid-19-cases-worldwide
}
