% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_npi_data.R
\name{load_npi_data}
\alias{load_npi_data}
\title{Get epidemiological data for 31 European countries}
\usage{
load_npi_data(end_date = Sys.Date())
}
\arguments{
\item{end_date}{The latest date taken into account. (Taken as the end date
for strategies where no end has been registered.)}
}
\description{
Get epidemiological data for 31 European countries
}
