% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumarise_estimation.R
\name{summarise_estimation}
\alias{summarise_estimation}
\title{Get 95\% credibility intervals for estimations}
\usage{
summarise_estimation(MCMC_file, npi_data, contact_data, age_data)
}
\description{
Get 95\% credibility intervals for estimations
}
