% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_contact_data.R
\name{load_contact_data}
\alias{load_contact_data}
\title{Get contact data matrix for a specific country}
\usage{
load_contact_data(country)
}
\arguments{
\item{country}{Character. The name of the country for which you want contact
data.}
}
\value{
A square (16 by 16) matrix containing the contact data between
the different age classes for a given country.
}
\description{
Get contact data matrix for a specific country
}
\examples{
load_contact_data("France")

}
\references{
Kiesha Prem, Alex R. Cook, Mark Jit, Projecting social contact
matrices in 152 countries using contact surveys and demographic data, PLoS
Comp. Biol. (2017), \ifelse{{text}{doi: 10.1371/journal.pcbi.1005697 (URL: https://doi.org/10.1371/journal.pcbi.1005697)}{\ifelse{{latex}{doi:\out{\\nobreakspace{}}\href{https://doi.org/10.1371/journal.pcbi.1005697}{10.1371\out{\\slash{}}journal.pcbi.1005697}}{doi: \href{https://doi.org/10.1371/journal.pcbi.1005697}{10.1371/journal.pcbi.1005697}}}}}
}
