% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumarise_estimation.R
\name{summarise_estimation}
\alias{summarise_estimation}
\title{Get 95\% credibility intervals for estimations}
\usage{
summarise_estimation(MCMC, npi_data)
}
\arguments{
\item{MCMC}{\code{mcmc} or \code{mcmc.list} object coda object}
}
\description{
Get 95\% credibility intervals for estimations
}
